+++ src/commands/index.ts
@@ -7,8 +7,9 @@ export const handleErrors = (handler: Handler): ((...args: any[]) => Promise<any
         try {
             await handler(...args);
         } catch (error) {
-            if (error.message) {
-                vscode.window.showErrorMessage(error.message);
+            const message = error instanceof Error ? error.message : undefined;
+            if (message) {
+                vscode.window.showErrorMessage(message);
             } else {
                 console.log(error);
                 vscode.window.showErrorMessage('An unexpected error occured');