+++ src/kafka-file/utils/async.ts
@@ -103,7 +103,7 @@ export class Throttler<T> {
 export class Delayer<T> {
 
 	public defaultDelay: number;
-	private timeout: NodeJS.Timer | null;
+	private timeout: ReturnType<typeof setTimeout> | null;
 	private completionPromise: Promise<T> | null;
 	private onResolve: ((value: T | PromiseLike<T> | undefined) => void) | null;
 	private task: ITask<T> | null;