+++ src/test/suite/kafka-file/languageservice/diagnosticsFakerJS.test.ts
@@ -8,29 +8,29 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words}}',
+            '{{lorem.words}}',
             []
         );
 
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words}}abcd{{random.words}}',
+            '{{lorem.words}}abcd{{lorem.words}}',
             []
         );
 
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words}}\n' +
-            '{{random.words}}',
+            '{{lorem.words}}\n' +
+            '{{lorem.words}}',
             []
         );
 
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words(5)}}',
+            '{{lorem.words(5)}}',
             []
         );
 
@@ -38,29 +38,21 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             'PRODUCER\n' +
             'topic:abcd\n' +
             `
-            {{address.zipCode}}
-            {{address.zipCodeByState}}
-            {{address.city}}
-            {{address.cityPrefix}}
-            {{address.citySuffix}}
-            {{address.cityName}}
-            {{address.streetName}}
-            {{address.streetAddress}}
-            {{address.streetSuffix}}
-            {{address.streetPrefix}}
-            {{address.secondaryAddress}}
-            {{address.county}}
-            {{address.country}}
-            {{address.countryCode}}
-            {{address.state}}
-            {{address.stateAbbr}}
-            {{address.latitude}}
-            {{address.longitude}}
-            {{address.direction}}
-            {{address.cardinalDirection}}
-            {{address.ordinalDirection}}
-            {{address.nearbyGPSCoordinate}}
-            {{address.timeZone}}
+            {{location.zipCode}}
+            {{location.city}}
+            {{location.streetAddress}}
+            {{location.secondaryAddress}}
+            {{location.county}}
+            {{location.country}}
+            {{location.countryCode}}
+            {{location.state}}
+            {{location.latitude}}
+            {{location.longitude}}
+            {{location.direction}}
+            {{location.cardinalDirection}}
+            {{location.ordinalDirection}}
+            {{location.nearbyGPSCoordinate}}
+            {{location.timeZone}}
             {{animal.dog}}
             {{animal.cat}}
             {{animal.snake}}
@@ -75,7 +67,6 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{animal.insect}}
             {{animal.rabbit}}
             {{animal.type}}
-            {{commerce.color}}
             {{commerce.department}}
             {{commerce.productName}}
             {{commerce.price}}
@@ -83,17 +74,15 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{commerce.productMaterial}}
             {{commerce.product}}
             {{commerce.productDescription}}
-            {{company.suffixes}}
-            {{company.companyName}}
-            {{company.companySuffix}}
+            {{company.name}}
             {{company.catchPhrase}}
-            {{company.bs}}
+            {{company.buzzPhrase}}
             {{company.catchPhraseAdjective}}
             {{company.catchPhraseDescriptor}}
             {{company.catchPhraseNoun}}
-            {{company.bsAdjective}}
-            {{company.bsBuzz}}
-            {{company.bsNoun}}
+            {{company.buzzAdjective}}
+            {{company.buzzVerb}}
+            {{company.buzzNoun}}
             {{database.column}}
             {{database.type}}
             {{database.collation}}
@@ -104,10 +93,10 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{date.soon}}
             {{date.month}}
             {{date.weekday}}
-            {{finance.account}}
+            {{finance.accountNumber}}
             {{finance.accountName}}
             {{finance.routingNumber}}
-            {{finance.mask}}
+            {{finance.maskedNumber}}
             {{finance.amount}}
             {{finance.transactionType}}
             {{finance.currencyCode}}
@@ -125,17 +114,16 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{git.commitEntry}}
             {{git.commitMessage}}
             {{git.commitSha}}
-            {{git.shortSha}}
             {{hacker.abbreviation}}
             {{hacker.adjective}}
             {{hacker.noun}}
             {{hacker.verb}}
             {{hacker.ingverb}}
             {{hacker.phrase}}
-            {{internet.avatar}}
+            {{image.avatar}}
             {{internet.email}}
             {{internet.exampleEmail}}
-            {{internet.userName}}
+            {{internet.username}}
             {{internet.protocol}}
             {{internet.httpMethod}}
             {{internet.url}}
@@ -159,32 +147,25 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{lorem.text}}
             {{lorem.lines}}
             {{music.genre}}
-            {{mersenne.rand}}
-            {{name.firstName}}
-            {{name.lastName}}
-            {{name.middleName}}
-            {{name.findName}}
-            {{name.jobTitle}}
-            {{name.gender}}
-            {{name.prefix}}
-            {{name.suffix}}
-            {{name.title}}
-            {{name.jobDescriptor}}
-            {{name.jobArea}}
-            {{name.jobType}}
-            {{phone.phoneNumber}}
-            {{phone.phoneNumberFormat}}
-            {{phone.phoneFormats}}
-            {{random.number}}
-            {{random.float}}
-            {{random.uuid}}
-            {{random.boolean}}
-            {{random.word}}
-            {{random.words}}
-            {{random.locale}}
-            {{random.alpha}}
-            {{random.alphaNumeric}}
-            {{random.hexaDecimal}}
+            {{person.firstName}}
+            {{person.lastName}}
+            {{person.middleName}}
+            {{person.fullName}}
+            {{person.jobTitle}}
+            {{person.gender}}
+            {{person.prefix}}
+            {{person.suffix}}
+            {{person.jobDescriptor}}
+            {{person.jobArea}}
+            {{person.jobType}}
+            {{phone.number}}
+            {{phone.imei}}
+            {{number.int}}
+            {{number.float}}
+            {{string.uuid}}
+            {{string.alpha}}
+            {{string.alphanumeric}}
+            {{string.hexadecimal}}
             {{system.fileName}}
             {{system.commonFileName}}
             {{system.mimeType}}
@@ -195,7 +176,6 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
             {{system.directoryPath}}
             {{system.filePath}}
             {{system.semver}}
-            {{time.recent}}
             {{vehicle.vehicle}}
             {{vehicle.manufacturer}}
             {{vehicle.model}}
@@ -215,12 +195,12 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            'random.words}}',
+            'lorem.words}}',
             [
                 diagnostic(
                     position(2, 0),
-                    position(2, 12),
-                    "FakerJS expression 'random.words' must be opened with '{{'",
+                    position(2, 11),
+                    "FakerJS expression 'lorem.words' must be opened with '{{'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -232,13 +212,13 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words}}\n' +
-            'random.words}}',
+            '{{lorem.words}}\n' +
+            'lorem.words}}',
             [
                 diagnostic(
-                    position(2, 16),
-                    position(3, 12),
-                    "FakerJS expression '\nrandom.words' must be opened with '{{'",
+                    position(2, 15),
+                    position(3, 11),
+                    "FakerJS expression '\nlorem.words' must be opened with '{{'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -250,12 +230,12 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.w{{ords}}',
+            '{{lorem.w{{ords}}',
             [
                 diagnostic(
-                    position(2, 10),
-                    position(2, 12),
-                    "Unexpected token '{{' in expression 'random.w{{ords'",
+                    position(2, 9),
+                    position(2, 11),
+                    "Unexpected token '{{' in expression 'lorem.w{{ords'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -267,12 +247,12 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words',
+            '{{lorem.words',
             [
                 diagnostic(
                     position(2, 2),
-                    position(2, 14),
-                    "FakerJS expression 'random.words' must be closed with '}}'",
+                    position(2, 13),
+                    "FakerJS expression 'lorem.words' must be closed with '}}'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -285,12 +265,12 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words{{',
+            '{{lorem.words{{',
             [
                 diagnostic(
                     position(2, 2),
-                    position(2, 16),
-                    "FakerJS expression 'random.words{{' must be closed with '}}'",
+                    position(2, 15),
+                    "FakerJS expression 'lorem.words{{' must be closed with '}}'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -302,13 +282,13 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words\n' +
-            '{{random.words}}',
+            '{{lorem.words\n' +
+            '{{lorem.words}}',
             [
                 diagnostic(
                     position(3, 0),
                     position(3, 2),
-                    "Unexpected token '{{' in expression 'random.words\n{{random.words'",
+                    "Unexpected token '{{' in expression 'lorem.words\n{{lorem.words'",
                     DiagnosticSeverity.Error
                 )
             ]
@@ -389,11 +369,11 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.BAD_METHOD}}',
+            '{{lorem.BAD_METHOD}}',
             [
                 diagnostic(
-                    position(2, 9),
-                    position(2, 19),
+                    position(2, 8),
+                    position(2, 18),
                     "Invalid method: 'BAD_METHOD'",
                     DiagnosticSeverity.Error
                 )
@@ -403,11 +383,11 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words()}}',
+            '{{lorem.words()}}',
             [
                 diagnostic(
-                    position(2, 9),
-                    position(2, 16),
+                    position(2, 8),
+                    position(2, 15),
                     "Invalid method: 'words()'",
                     DiagnosticSeverity.Error
                 )
@@ -417,11 +397,11 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words(}}',
+            '{{lorem.words(}}',
             [
                 diagnostic(
-                    position(2, 9),
-                    position(2, 15),
+                    position(2, 8),
+                    position(2, 14),
                     "Invalid method: 'words('",
                     DiagnosticSeverity.Error
                 )
@@ -431,7 +411,7 @@ suite("Kafka File PRODUCER FakerJS Diagnostics Test Suite", () => {
         await assertDiagnostics(
             'PRODUCER\n' +
             'topic:abcd\n' +
-            '{{random.words(5)}}',
+            '{{lorem.words(5)}}',
             []);
     });